backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-CORS
python-dotenv
requests
pandas
